[CL-94] chore: split data fetching stories into separate file
